Additions: The $search function allows users to paste a link to a supported site and gives them the option to add the manga to their list directly.
Improvements:
Better $updates information display. Updates are now all sent together in one block of text instead of multiple messages.

Fixes:
Fixed an issue that caused $updates to hang due to the aio_search_function not returning a value when the user did not have any new updates for a manga series.